# docs/openapi.yaml
openapi: 3.1.0
info:
  title: THE Travel Club API
  version: 1.0.0
  description: Enterprise contract for Travel Happier Everywhere aggregation endpoints.
servers:
  - url: /
    description: Same origin as the API host.
  - url: http://localhost:3000
    description: Local development default.
security:
  - ApiKeyAuth: []
paths:
...
  /v1/flights/search:
    get:
      summary: Search flight offers.
      parameters:
        - in: query
          name: from
          required: true
          description: Origin airport as a 3-letter IATA or 4-letter ICAO code (e.g. LAX, KLAX).
          schema: { type: string, minLength: 3, maxLength: 4, pattern: '^[A-Za-z]{3,4}$' }
        - in: query
          name: to
          required: true
          description: Destination airport as a 3-letter IATA or 4-letter ICAO code.
          schema: { type: string, minLength: 3, maxLength: 4, pattern: '^[A-Za-z]{3,4}$' }
        - in: query
          name: date
          description: Departure date (YYYY-MM-DD). Must be today or later.
          schema: { type: string, format: date }
        - in: query
          name: sort
          schema: { type: string, enum: [price, score] }
        - in: query
          name: limit
          description: Maximum number of offers to return (1-50).
          schema: { type: integer, minimum: 1, maximum: 50 }
      responses:
        '200': { description: Ranked offers. }
        '400': { description: Invalid query. }
        '401': { description: Authentication required. }
        '403': { description: Forbidden. }
        '429':
          description: Rate limit exceeded.
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema: { type: integer }
  /v1/flights/calendar:
    get:
      summary: Cheapest comparable price per day across a date window (flexible dates).
      parameters:
        - in: query
          name: from
          required: true
          description: Origin airport as a 3-letter IATA or 4-letter ICAO code.
          schema: { type: string, minLength: 3, maxLength: 4, pattern: '^[A-Za-z]{3,4}$' }
        - in: query
          name: to
          required: true
          description: Destination airport as a 3-letter IATA or 4-letter ICAO code.
          schema: { type: string, minLength: 3, maxLength: 4, pattern: '^[A-Za-z]{3,4}$' }
        - in: query
          name: date
          required: true
          description: Center date (YYYY-MM-DD). Must be today or later.
          schema: { type: string, format: date }
        - in: query
          name: flex
          description: Days to search on each side of the center date (1-7, default 3; clamped, not rejected).
          schema: { type: integer, minimum: 1, maximum: 7 }
      responses:
        '200': { description: Per-day cheapest totals plus the cheapest date in the window. }
        '400': { description: Invalid query or missing center date. }
        '429':
          description: Rate limit exceeded.
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema: { type: integer }
  /v1/hotels/search:
    get:
      summary: Search hotel offers.
      parameters:
        - in: query
          name: city
          required: true
          schema: { type: string, minLength: 1, maxLength: 120 }
        - in: query
          name: cityCode
          description: Optional 3-letter IATA/destination code used by real hotel providers (e.g. Hotelbeds).
          schema: { type: string, minLength: 3, maxLength: 3 }
        - in: query
          name: checkin
          description: Check-in date (YYYY-MM-DD). Must be today or later.
          schema: { type: string, format: date }
        - in: query
          name: checkout
          description: Check-out date (YYYY-MM-DD). Must be after check-in.
          schema: { type: string, format: date }
        - in: query
          name: sort
          schema: { type: string, enum: [price, score] }
        - in: query
          name: limit
          description: Maximum number of offers to return (1-50).
          schema: { type: integer, minimum: 1, maximum: 50 }
      responses:
        '200': { description: Ranked offers. }
        '400': { description: Invalid query. }
        '429':
          description: Rate limit exceeded.
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema: { type: integer }
  /v1/cars/search:
    get:
      summary: Search car rental offers.
      parameters:
        - in: query
          name: city
          required: true
          schema: { type: string, minLength: 1, maxLength: 120 }
        - in: query
          name: date
          description: Rental date (YYYY-MM-DD). Must be today or later.
          schema: { type: string, format: date }
        - in: query
          name: sort
          schema: { type: string, enum: [price, score] }
        - in: query
          name: limit
          description: Maximum number of offers to return (1-50).
          schema: { type: integer, minimum: 1, maximum: 50 }
      responses:
        '200': { description: Ranked offers. }
        '400': { description: Invalid query. }
        '429':
          description: Rate limit exceeded.
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema: { type: integer }
  /v1/prices/history:
    get:
      summary: Recent cheapest-price history for a search (30-day window).
      parameters:
        - in: query
          name: type
          required: true
          schema: { type: string, enum: [flights, hotels, cars] }
        - in: query
          name: from
          description: Origin airport code (required when type=flights).
          schema: { type: string, minLength: 3, maxLength: 4 }
        - in: query
          name: to
          description: Destination airport code (required when type=flights).
          schema: { type: string, minLength: 3, maxLength: 4 }
        - in: query
          name: city
          description: City (required when type=hotels or type=cars).
          schema: { type: string, minLength: 1, maxLength: 120 }
        - in: query
          name: checkin
          description: Optional stay start; hotels aggregate per stay length.
          schema: { type: string, format: date }
        - in: query
          name: checkout
          schema: { type: string, format: date }
      responses:
        '200': { description: Samples, average, lowest, latest, and a recent series. }
        '400': { description: Invalid or missing parameters. }
        '404': { description: Price history is disabled. }
  /v1/alerts:
    get:
      summary: List the caller's price alerts / saved searches (owner-scoped by API key).
      responses:
        '200': { description: The caller's watches. }
    post:
      summary: Create a price alert / saved search. A watch with a threshold notifies when the cheapest total drops to it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string, enum: [flights, hotels, cars] }
                from: { type: string, description: 'Flights origin code.' }
                to: { type: string, description: 'Flights destination code.' }
                date: { type: string, format: date }
                city: { type: string, description: 'Hotels/cars city.' }
                checkin: { type: string, format: date }
                checkout: { type: string, format: date }
                threshold: { type: number, minimum: 0, description: 'Notify when the cheapest total is at or below this. Omit for a plain saved search.' }
                currency: { type: string }
                notifyUrl: { type: string, description: 'Optional webhook URL (delivery is off unless ALERTS_WEBHOOKS_ENABLED; SSRF-guarded). Never echoed back.' }
      responses:
        '201': { description: The created watch (the notify URL is never returned; only notifyConfigured). }
        '400': { description: Invalid body or query. }
        '404': { description: Alerts are disabled. }
    delete:
      summary: Delete one of the caller's alerts by id.
      parameters:
        - in: query
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Deleted. }
        '400': { description: Missing id. }
        '404': { description: Alert not found (or not owned by the caller). }
  /v1/auth/signup:
    post:
      security: []
      summary: Create a member account and start a session (sets an HttpOnly session cookie).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8, description: 'At least 8 characters.' }
      responses:
        '201':
          description: The created member (public shape; never includes the password hash). A tc_session cookie is set.
          headers:
            Set-Cookie:
              description: HttpOnly, SameSite=Lax session cookie (Secure in production).
              schema: { type: string }
        '400': { description: Invalid email or weak password. }
        '404': { description: Accounts are not enabled on this deployment. }
        '409': { description: An account with that email already exists. }
  /v1/auth/login:
    post:
      security: []
      summary: Sign in and start a session (sets an HttpOnly session cookie).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: The member (public shape). A tc_session cookie is set.
          headers:
            Set-Cookie:
              description: HttpOnly, SameSite=Lax session cookie (Secure in production).
              schema: { type: string }
        '401': { description: Invalid email or password. }
        '404': { description: Accounts are not enabled on this deployment. }
  /v1/auth/logout:
    post:
      security: []
      summary: End the current session (clears the session cookie).
      responses:
        '200':
          description: Signed out; the tc_session cookie is cleared.
        '404': { description: Accounts are not enabled on this deployment. }
  /v1/me:
    get:
      security: []
      summary: The signed-in member (tier, benefits, loyalty balance). Requires a session cookie.
      responses:
        '200': { description: The member profile (public shape). }
        '401': { description: Not signed in. }
        '404': { description: Accounts are not enabled on this deployment. }
  /v1/orders:
    get:
      security: []
      summary: List the caller's booking orders (owner-scoped by session or API key).
      responses:
        '200': { description: The caller's orders. }
        '404': { description: Booking is not enabled on this deployment. }
    post:
      security: []
      summary: Book an offer via the aggregator (merchant of record). Adds a tier-aware service fee.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, offer, passengers, contact]
              properties:
                type: { type: string, enum: [flights, hotels, cars] }
                offer: { type: object, description: 'A selected offer snapshot (id, price, and provider detail from a search response).' }
                passengers:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [givenName, familyName]
                    properties:
                      givenName: { type: string }
                      familyName: { type: string }
                contact:
                  type: object
                  required: [email]
                  properties:
                    email: { type: string, format: email }
                    phone: { type: string }
      responses:
        '201': { description: The confirmed order (confirmation, provider reference, and an all-in total including the service fee). }
        '400': { description: Invalid body, offer, passengers, or contact. }
        '404': { description: Booking is not enabled on this deployment. }
        '502': { description: The aggregator could not complete the booking. }
...
  /v1/billing:
    get:
      security: []
      summary: The signed-in member's subscription status and gateway mode. Requires a session.
      responses:
        '200': { description: Current tier, subscription status, tier, period end, and whether the gateway is live. }
        '401': { description: Not signed in. }
        '404': { description: Billing is not enabled on this deployment. }
  /v1/billing/subscribe:
    post:
      security: []
      summary: Subscribe the signed-in member to a paid tier (upgrades their membership).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tier]
              properties:
                tier: { type: string, enum: [silver, gold] }
      responses:
        '200': { description: The upgraded member and the subscription summary. }
        '400': { description: Not a paid tier. }
        '401': { description: Not signed in. }
        '404': { description: Billing is not enabled. }
  /v1/billing/cancel:
    post:
      security: []
      summary: Cancel the signed-in member's subscription (downgrades to free).
      responses:
        '200': { description: The downgraded member and the cancelled subscription. }
        '400': { description: No active subscription. }
        '401': { description: Not signed in. }
        '404': { description: Billing is not enabled. }
  /v1/billing/webhook:
    post:
      security: []
      summary: Gateway webhook endpoint. Signature-verified when a webhook secret is configured.
      responses:
        '200': { description: The event was processed (applied or ignored). }
        '400': { description: Malformed webhook payload. }
        '401': { description: Invalid webhook signature. }
        '404': { description: Billing is not enabled. }
  /v1/airport/info:
    get:
      summary: Look up airport information.
      parameters:
        - in: query
          name: code
          required: true
          description: Airport as a 3-letter IATA or 4-letter ICAO code.
          schema: { type: string, minLength: 3, maxLength: 4, pattern: '^[A-Za-z]{3,4}$' }
      responses:
        '200': { description: Airport information. }
        '400': { description: Invalid query. }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key